updated code and new functional tool code added.
